requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
distlib==0.3.7
filelock==3.12.2
idna==3.6
importlib-metadata==6.7.0
oauthlib==3.2.2
platformdirs==4.0.0
python-dotenv==0.21.1
requests==2.31.0
requests-oauthlib==1.3.1
tweepy==4.0.0
typing_extensions==4.7.1
urllib3<2.0
virtualenv==20.24.7
zipp==3.15.0
